+++ oidc-rs-spring-boot-web-config/src/main/java/org/entur/auth/spring/config/server/IssuerAuthenticationManagerResolver.java
@@ -36,10 +36,7 @@ public final class IssuerAuthenticationManagerResolver
             @NonNull List<JWKSourceWithIssuer> remoteJWKSets,
             @NonNull EnturAuthProperties enturAuthProperties,
             @NonNull TenantJwtGrantedAuthoritiesConverter authoritiesConverter,
-            @NonNull
-                    HealthReportListener<
-                                    JWKSetSourceWithHealthStatusReporting<SecurityContext>, SecurityContext>
-                            healthReportListener) {
+            HealthReportListener<JWKSetSourceWithHealthStatusReporting<SecurityContext>, SecurityContext> healthReportListener) {
 
         this.authenticationManagers = authenticationManagers;
         this.remoteJWKSets = new ArrayList<>(remoteJWKSets);