+++ oidc-rs-spring-boot-web-test/src/test/java/org/entur/auth/spring/test/other/AuthorityWebMvcTest.java
@@ -9,7 +9,6 @@ import org.entur.auth.junit.tenant.TenantJsonWebToken;
 import org.entur.auth.spring.application.GreetingController;
 import org.entur.auth.spring.config.ConfigAuthManagerResolverAutoConfiguration;
 import org.entur.auth.spring.config.ConfigAuthProvidersAutoConfiguration;
-import org.entur.auth.spring.config.ConfigJwksHealthIndicatorAutoConfiguration;
 import org.entur.auth.spring.config.ConfigResourceServerAutoConfiguration;
 import org.entur.auth.spring.web.ResourceServerAutoConfiguration;
 import org.entur.auth.spring.web.ResourceServerDefaultAutoConfiguration;
@@ -31,18 +30,14 @@ import org.springframework.test.web.servlet.MockMvc;
         })
 @ExtendWith({
     TenantJsonWebToken.class
-}) // Needed to define MOCKAUTHSERVER_PORT and support parameter to be resolved
+})
 @WebMvcTest({GreetingController.class})
 @ImportAutoConfiguration({
-    ResourceServerAutoConfiguration.class, // Configure default SecurityFilterChain
-    ResourceServerDefaultAutoConfiguration.class, // Configure default Spring Security behavior
-    ConfigAuthProvidersAutoConfiguration.class, // Provide predefined tenants environment
-    ConfigAuthManagerResolverAutoConfiguration.class, // Needed to use tenants.environment
-    ConfigResourceServerAutoConfiguration.class, // Configure authorization, cors, mdc, issuers
-    ConfigJwksHealthIndicatorAutoConfiguration
-            .class // Needed to get management endpoint jwksState working
-    // ConfigExternalPropertyAutoConfiguration.class      // Needed if you read issuers from property
-    // file
+    ResourceServerAutoConfiguration.class,
+    ResourceServerDefaultAutoConfiguration.class,
+    ConfigAuthProvidersAutoConfiguration.class,
+    ConfigAuthManagerResolverAutoConfiguration.class,
+    ConfigResourceServerAutoConfiguration.class
 })
 public class AuthorityWebMvcTest {
     @Autowired private MockMvc mockMvc;